+++ proxy/manager_test.go
@@ -29,47 +29,21 @@ func Test_Manager_SetRoutes_returnsErrorIfGetCertificateFails(t *testing.T) {
 		err: fmt.Errorf("ruh roh"),
 	}
 
-	manager := NewManager(nil, map[string]CertificateProvider{"fake": certManager}, "fake")
+	manager := NewManager(map[string]CertificateProvider{"fake": certManager}, "fake")
 	err := manager.SetRoutes([]*Route{{
 		Domains: []string{"example.com"},
 	}})
 	assert.Error(t, err)
 	assert.Equal(t, "example.com", certManager.subject)
-	assert.Equal(t, []string(nil), certManager.altNames)
+	assert.Equal(t, []string{}, certManager.altNames)
 }
 
 func Test_Manager_SetRoutes_returnsErrorIfDomainisInvalid(t *testing.T) {
-	manager := NewManager(nil, nil, "")
+	manager := NewManager(nil, "")
 	err := manager.SetRoutes([]*Route{{
 		Domains: []string{"example..com"},
 	}})
 	assert.Error(t, err)
-}
-
-func Test_Manager_SetRoutes_requestsWildcardCertificateIfMatching(t *testing.T) {
-	certManager := &fakeCertManager{
-		err: fmt.Errorf("ruh roh"),
-	}
-
-	manager := NewManager([]string{".example.com"}, map[string]CertificateProvider{"fake": certManager}, "fake")
-	_ = manager.SetRoutes([]*Route{{
-		Domains: []string{"test.example.com"},
-	}})
-	assert.Equal(t, "*.example.com", certManager.subject)
-	assert.EqualValues(t, []string(nil), certManager.altNames)
-}
-
-func Test_Manager_SetRoutes_translatesAltNamesToWildcardsIfMatching(t *testing.T) {
-	certManager := &fakeCertManager{
-		err: fmt.Errorf("ruh roh"),
-	}
-
-	manager := NewManager([]string{".example.com"}, map[string]CertificateProvider{"fake": certManager}, "fake")
-	_ = manager.SetRoutes([]*Route{{
-		Domains: []string{"test.deep.example.com", "test.example.com", "example.com"},
-	}})
-	assert.Equal(t, "test.deep.example.com", certManager.subject)
-	assert.EqualValues(t, []string{"*.example.com", "example.com"}, certManager.altNames)
 }
 
 func Test_Manager_RouteForDomain_returnsNullIfNoRouteFound(t *testing.T) {
@@ -77,7 +51,7 @@ func Test_Manager_RouteForDomain_returnsNullIfNoRouteFound(t *testing.T) {
 		err: fmt.Errorf("ruh roh"),
 	}
 
-	manager := NewManager(nil, map[string]CertificateProvider{"fake": certManager}, "fake")
+	manager := NewManager(map[string]CertificateProvider{"fake": certManager}, "fake")
 	res := manager.RouteForDomain("example.com")
 	assert.Nil(t, res)
 }
@@ -87,7 +61,7 @@ func Test_Manager_RouteForDomain_returnsCertificateForDomain(t *testing.T) {
 		certificate: dummyCert,
 	}
 
-	manager := NewManager(nil, map[string]CertificateProvider{"fake": certManager}, "fake")
+	manager := NewManager(map[string]CertificateProvider{"fake": certManager}, "fake")
 	route := &Route{
 		Domains: []string{"test.deep.example.com", "test.example.com", "example.com"},
 	}
@@ -103,7 +77,7 @@ func Test_Manager_CertificateForClient_returnsNullIfNoRouteFound(t *testing.T) {
 		err: fmt.Errorf("ruh roh"),
 	}
 
-	manager := NewManager(nil, map[string]CertificateProvider{"fake": certManager}, "fake")
+	manager := NewManager(map[string]CertificateProvider{"fake": certManager}, "fake")
 	res, err := manager.CertificateForClient(&tls.ClientHelloInfo{ServerName: "example.com"})
 	assert.Nil(t, res)
 	assert.Nil(t, err)
@@ -114,7 +88,7 @@ func Test_Manager_CertificateForClient_returnsCertificateForDomain(t *testing.T)
 		certificate: dummyCert,
 	}
 
-	manager := NewManager(nil, map[string]CertificateProvider{"fake": certManager}, "fake")
+	manager := NewManager(map[string]CertificateProvider{"fake": certManager}, "fake")
 	route := &Route{
 		Domains: []string{"test.deep.example.com", "test.example.com", "example.com"},
 	}
@@ -138,7 +112,7 @@ func Test_Manager_setsCertificateOnRoutes(t *testing.T) {
 		certificate: dummyCert,
 	}
 
-	manager := NewManager(nil, map[string]CertificateProvider{"fake": certManager}, "fake")
+	manager := NewManager(map[string]CertificateProvider{"fake": certManager}, "fake")
 	_ = manager.SetRoutes([]*Route{{
 		Domains: []string{"test.deep.example.com", "test.example.com", "example.com"},
 	}})
@@ -153,7 +127,7 @@ func Test_Manager_SetRoutes_removesPreviousRoutes(t *testing.T) {
 		certificate: dummyCert,
 	}
 
-	manager := NewManager(nil, map[string]CertificateProvider{"fake": certManager}, "fake")
+	manager := NewManager(map[string]CertificateProvider{"fake": certManager}, "fake")
 	_ = manager.SetRoutes([]*Route{{
 		Domains: []string{"test.deep.example.com", "test.example.com", "example.com"},
 	}})
@@ -169,7 +143,7 @@ func Test_Manager_CheckCertificates_returnsErrorIfGetCertificateFails(t *testing
 		certificate: dummyCert,
 	}
 
-	manager := NewManager(nil, map[string]CertificateProvider{"fake": certManager}, "fake")
+	manager := NewManager(map[string]CertificateProvider{"fake": certManager}, "fake")
 	_ = manager.SetRoutes([]*Route{{
 		Domains: []string{"test.deep.example.com", "test.example.com", "example.com"},
 	}})
@@ -183,7 +157,7 @@ func Test_Manager_CheckCertificates_usesSupplierSpecifiedByRoute(t *testing.T) {
 	certManager1 := &fakeCertManager{}
 	certManager2 := &fakeCertManager{}
 
-	manager := NewManager(nil, map[string]CertificateProvider{"f1": certManager1, "f2": certManager2}, "f1")
+	manager := NewManager(map[string]CertificateProvider{"f1": certManager1, "f2": certManager2}, "f1")
 	_ = manager.SetRoutes([]*Route{{
 		Domains:  []string{"test.deep.example.com", "test.example.com", "example.com"},
 		Provider: "f2",
@@ -197,7 +171,7 @@ func Test_Manager_CheckCertificates_fallsBackToDefaultSupplier(t *testing.T) {
 	certManager1 := &fakeCertManager{}
 	certManager2 := &fakeCertManager{}
 
-	manager := NewManager(nil, map[string]CertificateProvider{"f1": certManager1, "f2": certManager2}, "f2")
+	manager := NewManager(map[string]CertificateProvider{"f1": certManager1, "f2": certManager2}, "f2")
 	_ = manager.SetRoutes([]*Route{{
 		Domains: []string{"test.deep.example.com", "test.example.com", "example.com"},
 	}})
@@ -210,7 +184,7 @@ func Test_Manager_CheckCertificates_errorsIfProviderNotFound(t *testing.T) {
 	certManager1 := &fakeCertManager{}
 	certManager2 := &fakeCertManager{}
 
-	manager := NewManager(nil, map[string]CertificateProvider{"f1": certManager1, "f2": certManager2}, "f2")
+	manager := NewManager(map[string]CertificateProvider{"f1": certManager1, "f2": certManager2}, "f2")
 	err := manager.SetRoutes([]*Route{{
 		Domains:  []string{"test.deep.example.com", "test.example.com", "example.com"},
 		Provider: "c3p0",
@@ -224,7 +198,7 @@ func Test_Manager_CheckCertificates_updatesAllCertificates(t *testing.T) {
 		certificate: dummyCert,
 	}
 
-	manager := NewManager(nil, map[string]CertificateProvider{"fake": certManager}, "fake")
+	manager := NewManager(map[string]CertificateProvider{"fake": certManager}, "fake")
 	_ = manager.SetRoutes([]*Route{{
 		Domains: []string{"test.deep.example.com", "test.example.com", "example.com"},
 	}, {