+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Post v0.1.0
 
+## Features
+
+- Centauri is now capable of generating self-signed certificates
+  for routes, instead of obtaining them via ACME. This is controlled
+  on a per-route basis by using the new `provider selfsigned` directive.
+  ([issue #15](https://github.com/csmith/centauri/issues/15))
+
 ## Other changes
 
 - Now requires Go 1.18 to build.